part 3 v 1.1

+++ src/main/java/com/pismennaya/shop/models/Product.java
@@ -41,7 +41,7 @@ public class Product implements CommonEntity<Long>  {
     private int power;
     private String material;
     private int steam_suply;
-    private float warranty;
+    private int warranty;
     private String description;
 
     public Product() {}
